Added throwError
This function is a complete error handler designed for debug/better error tracing and reduce several lines from your code. This code also provides a tiny tool that prevents some errors (for example, if it throws an error and it gets deleted before the autodelete, it'll also check if the bot can send the error message. (More security? Of course!).

Author's note: This is my favourite function, I use it in all pieces for both debug and error catching.