[WIP] Enhancements (#56)
* Fixed an issue when trying to add bots in the friendlist

Even tho the Endpoint doesn't seem to work

* Let's use request-promise-native

Tested!

* Use request-promise-native

* Added catfacts

More fun commands ^-^

* Add files via upload

* Added softban, fixed ban

* Added splitText to requiredFuncs

* Added cooldown v2

* Fixed didyoumean

> Checks if the bot is configured as a selfbot, to prevent some issues (and also make it run only for the owner)
> Filters the commands by permLevel
> Parse both prefix and command with `parseCommand`
> Add autodelete (if the message was sent)

* Added nomention

* Woops.

* Added throwError

This function is a complete error handler designed for debug/better error tracing and reduce several lines from your code. This code also provides a tiny tool that prevents some errors (for example, if it throws an error and it gets deleted before the autodelete, it'll also check if the bot can send the error message. (More security? Of course!).

Author's note: This is my favourite function, I use it in all pieces for both debug and error catching.

* Show channel and guild

* Made the pieces downloadable again

Fixes `require is not defined` in several pieces.

* Woops! (Re-sync)