Update purge command
### Changes Proposed in this Pull Request
- Add extended help
- selfbot: false, so I think, no need for `if (client.config.selfbot)
- When user mentioned, it fetches messages up to 100 (discord.js fetchMessages upper limit), then filter msg with author.id = mentioned user, then delete msg from that user X amount.
- When user not mentioned, it simply delete all last X amount of msg.

### (If Applicable) It will fix issues
Not really fixed the issue, but the real issue is, with this code, when I fetch msg (max is 100), then filter it from mentioned user, then delete all msg from that user by X amount (let's say, 100), it will delete all 100 msg from that user in the fetched msg. But, what if in that fetched msg not reached 100 msg of that user? It will return the error ¯\_(ツ)_/¯
Well, this is what I can do, I limit the amount from 2 to 25.

Hope there's a better solution. Thank you.